requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
boto3==1.37.11
PyYAML==6.0.2
pydantic==2.10.6
